+++ src/utils.js
@@ -7,9 +7,42 @@
  */
 
 import fs from 'fs-plus';
+import os from 'os';
 import path from 'path';
+import qs from 'querystringify';
 import vscode from 'vscode';
 
+
+export async function notifyError(title, err) {
+  const description = err.stack || err.toString();
+  const action = 'Report a problem';
+  const selected = await vscode.window.showErrorMessage(description, action);
+  if (selected === action) {
+    const ghbody = `# Description of problem
+Leave a comment...
+
+BEFORE SUBMITTING, PLEASE SEARCH FOR DUPLICATES IN
+- https://github.com/platformio/platformio-vscode-ide/issues
+
+# Configuration
+
+VSCode: ${vscode.version}
+PIO IDE: v${getIDEVersion()}
+System: ${os.type()}, ${os.release()}, ${os.arch()}
+
+# Exception
+\`\`\`
+${description}
+\`\`\`
+`;
+    vscode.commands.executeCommand(
+      'vscode.open',
+      vscode.Uri.parse(`https://github.com/platformio/platformio-vscode-ide/issues/new?${qs.stringify(
+        { title: encodeURIComponent(title), body: encodeURIComponent(ghbody) })}`)
+    );
+  }
+  console.error(err);
+}
 
 export function getIDEManifest() {
   return vscode.extensions.getExtension('platformio.platformio-ide').packageJSON;