+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release Notes
 
+## 0.16.0 (2018-06-20)
+
+**Requires VSCode 1.24 or above**
+
+* [Custom Tasks](http://docs.platformio.org/en/latest/ide/vscode.html#custom-tasks) (issue [#89](https://github.com/platformio/platformio-vscode-ide/issues/89))
+* Automatically close Serial Port Monitor before uploading/testing (issue [#49](https://github.com/platformio/platformio-vscode-ide/issues/49))
+* Added new configuration option `autoCloseSerialMonitor`, which is set to `true` by default
+* Added "Report a problem" action/button when error occurs
+* Improved PIO Core installer using `pip` as Python module
+
 ## 0.15.2 (2018-05-30)
 
 * Reverted back an order of PlatformIO Toolbar (issue [#114](https://github.com/platformio/platformio-vscode-ide/issues/114))
@@ -12,7 +22,7 @@
 
 ## 0.15.0 (2018-05-08)
 
-**Requires VSCode 1.23.0 or above**
+**Requires VSCode 1.23 or above**
 
 * New UI for [PIO Unified Debugger](http://docs.platformio.org/page/plus/debugging.html):
   - Conditional Breakpoints